Keep hierarchy when updating store with `use()` function
Previously, when store were updated with `Provider.use()` function
it were moved as last in configuration hierarchy. This caused
that usually some other configuration were overriding updated
configurations and the updates never get applied.
Updated `use()` function to not remove/add the store, but just
re-define it so the store key keep the same position in the
`this.stores` variable.