Fix insecure use of AES-256-CTR cipher (#296)
Using the same IV twice with a counter mode cipher breaks the security
of the cipher.  Switch to AES-256-CBC.

I made the cipher nonconfigurable because it seems more likely than not
that people will get it wrong otherwise.

This commit also drops the dependency on the secure-keys module, which
appears to be unmaintained and didn't really live up to its name.